refactor(#1235): replace DateTimeGranularity with Timescale
