+++ generator/src/main/java/com/scottlogic/deg/generator/restrictions/linear/NumericRestrictions.java
@@ -16,6 +16,8 @@
 
 package com.scottlogic.deg.generator.restrictions.linear;
 
+import com.scottlogic.deg.common.profile.constraintdetail.Granularity;
+
 import java.math.BigDecimal;
 import java.util.Objects;
