+++ generator/src/test/java/com/scottlogic/deg/generator/restrictions/DateTimeRestrictionsTests.java
@@ -39,7 +39,7 @@ class DateTimeRestrictionsTests {
     private static final OffsetDateTime MAX = granularToMillis(Defaults.ISO_MAX_DATE);
 
     private static OffsetDateTime granularToMillis(OffsetDateTime date) {
-        return Timescale.MILLIS.getGranularityFunction().apply(date);
+        return Timescale.MILLIS.trimToGranularity(date);
     }
 
     @Test
